Fix signing hash of batch and shutter txs
For Shutter txs, the signing hash excluded the L1BlockNumber field. This
was missed when we added the field to the transaction type.

For Batch txs, the signing hash included the decryption key. This makes
it impossible for the collator to commit to a batch which has to be done
before the decryption key is generated.